refactor(graduate): ui 디렉토리 네이밍 변경 components -> ui
