+++ apps/graduate/src/admin/pages/all/ui/AllManagementPage.tsx
@@ -1,4 +1,4 @@
-import { Header, Pagination } from '~/shared/components';
+import { Header, Pagination } from '~/shared/ui/index.ts';
 
 import {
   LOADING_TEXT,
@@ -9,7 +9,7 @@ import { extractPeriodData } from '../utils';
 import UserDetailModal from './UserDetailModal.tsx';
 import { useAllManagement } from '../model/useAllManagement.ts';
 
-import { DataTable, Toolbar } from '~/admin/shared/components';
+import { DataTable, Toolbar } from '~/admin/shared/ui/index.ts';
 import * as style from '~/admin/shared/styles/adminPage.css';
 
 export default function AllManagementPage() {
