+++ apps/graduate/src/admin/pages/all/ui/UserDetailModal.tsx
@@ -1,7 +1,7 @@
 import { useNavigate } from '@tanstack/react-router';
 import { Descriptions, Modal, Spin, Table } from 'antd';
 
-import { Header } from '~/shared/components';
+import { Header } from '~/shared/ui';
 
 import type { PeriodData, StageData } from '../types/allManagement';
 import {
@@ -11,7 +11,7 @@ import {
 } from '../utils';
 
 import { useStudentDetail } from '~/admin/features/studentDetail';
-import { Container } from '~/admin/shared/components/Container';
+import { Container } from '~/admin/shared/ui/Container';
 import { vars } from '~/vars.css';
 
 interface UserDetailModalProps {
