+++ apps/graduate/src/admin/shared/ui/TextEditor/types.ts
