+++ apps/graduate/src/admin/widgets/Table/ui/Table.tsx
@@ -3,7 +3,7 @@ import { GraduationUserSummary } from '~/admin/entities/graduation-users/api';
 import { useFetchGraduationUsers } from '~/admin/entities/graduation-users/model';
 import { certColumns } from '~/admin/pages/certification/constants/certColumns';
 import { thesisColumns } from '~/admin/pages/thesis/constants/thesisColumns';
-import { DataTable, Toolbar } from '~/admin/shared/components';
+import { DataTable, Toolbar } from '~/admin/shared/ui';
 import * as style from '~/admin/shared/styles/adminPage.css';
 import {
   useAdminSelection,
