+++ apps/graduate/src/client/pages/status/ui/StatusPage.tsx
@@ -4,7 +4,7 @@ import { useState, useMemo } from 'react';
 import { Calendar, dayjsLocalizer } from 'react-big-calendar';
 import 'react-big-calendar/lib/css/react-big-calendar.css';
 
-import { LoadingState, EmptyState } from '~/shared/components';
+import { LoadingState, EmptyState } from '~/shared/ui';
 import {
   SCHEDULE_COLOR_PALETTE,
   LOADING_MESSAGES,
