+++ apps/graduate/src/client/pages/thesis/ui/ThesisPage.tsx
@@ -1,6 +1,6 @@
 import { useSearch } from '@tanstack/react-router';
 
-import { Section } from '~/shared/components';
+import { Section } from '~/shared/ui';
 import * as styles from '~/client/shared/styles/SubmissionPage.css';
 
 import { ThesisCollapse } from './ThesisCollapse';
