+++ apps/graduate/src/shared/ui/EmptyState/EmptyState.css.ts
