+++ apps/graduate/src/shared/ui/EmptyState/EmptyState.tsx
