+++ apps/graduate/src/shared/ui/EmptyState/index.ts
