+++ apps/graduate/src/shared/ui/ErrorState/ErrorState.css.ts
