+++ apps/graduate/src/shared/ui/ErrorState/ErrorState.tsx
