+++ apps/graduate/src/shared/ui/ErrorState/index.ts
