+++ apps/graduate/src/shared/ui/LoadingState/LoadingState.css.ts
