+++ apps/graduate/src/shared/ui/LoadingState/LoadingState.tsx
