+++ apps/graduate/src/shared/ui/LoadingState/index.ts
