+++ apps/graduate/src/shared/ui/Section/Section.css.ts
