+++ apps/graduate/src/shared/ui/Section/Section.tsx
