+++ apps/graduate/src/shared/ui/Section/index.ts
